+++ tests/behavior/behavior.rs
@@ -78,23 +78,7 @@ macro_rules! behavior_tests {
             blocking_behavior_test!(
                 $service,
 
-                test_blocking_create_file,
-                test_blocking_create_file_existing,
-                test_blocking_create_file_with_special_chars,
-                test_blocking_create_dir,
-                test_blocking_create_dir_exising,
-                test_blocking_write,
-                test_blocking_write_with_dir_path,
-                test_blocking_write_with_special_chars,
-                test_blocking_stat,
-                test_blocking_stat_dir,
-                test_blocking_stat_with_special_chars,
-                test_blocking_stat_not_exist,
-                test_blocking_read_full,
-                test_blocking_read_range,
-                test_blocking_read_not_exist,
                 test_blocking_list_dir,
-                test_blocking_delete,
             );
         )*
     };
@@ -263,284 +247,6 @@ async fn test_object_name(op: Operator) -> Result<()> {
 
     let o = op.object(&path);
     assert_eq!(o.name(), format!("{name}/"));
-
-    Ok(())
-}
-
-/// Create file with file path should succeed.
-fn test_blocking_create_file(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-
-    let o = op.object(&path);
-
-    o.blocking_create()?;
-
-    let meta = o.blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::FILE);
-    assert_eq!(meta.content_length(), 0);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Create file on existing file path should succeed.
-fn test_blocking_create_file_existing(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-
-    let o = op.object(&path);
-
-    o.blocking_create()?;
-
-    o.blocking_create()?;
-
-    let meta = o.blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::FILE);
-    assert_eq!(meta.content_length(), 0);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Create file with special chars should succeed.
-fn test_blocking_create_file_with_special_chars(op: Operator) -> Result<()> {
-    let path = format!("{} !@#$%^&*()_+-=;'><,?.txt", uuid::Uuid::new_v4());
-
-    let o = op.object(&path);
-    debug!("{o:?}");
-
-    o.blocking_create()?;
-
-    let meta = o.blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::FILE);
-    assert_eq!(meta.content_length(), 0);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Create dir with dir path should succeed.
-fn test_blocking_create_dir(op: Operator) -> Result<()> {
-    let path = format!("{}/", uuid::Uuid::new_v4());
-
-    let o = op.object(&path);
-
-    o.blocking_create()?;
-
-    let meta = o.blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::DIR);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Create dir on existing dir should succeed.
-fn test_blocking_create_dir_exising(op: Operator) -> Result<()> {
-    let path = format!("{}/", uuid::Uuid::new_v4());
-
-    let o = op.object(&path);
-
-    o.blocking_create()?;
-
-    o.blocking_create()?;
-
-    let meta = o.blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::DIR);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Write a single file and test with stat.
-fn test_blocking_write(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-
-    op.object(&path).blocking_write(content)?;
-
-    let meta = op
-        .object(&path)
-        .blocking_metadata()
-        .expect("stat must succeed");
-    assert_eq!(meta.content_length(), size as u64);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Write file with dir path should return an error
-fn test_blocking_write_with_dir_path(op: Operator) -> Result<()> {
-    let path = format!("{}/", uuid::Uuid::new_v4());
-    let (content, _) = gen_bytes();
-
-    let result = op.object(&path).blocking_write(content);
-    assert!(result.is_err());
-    assert!(result.unwrap_err().to_string().contains("Is a directory"));
-
-    Ok(())
-}
-
-/// Write a single file with special chars should succeed.
-fn test_blocking_write_with_special_chars(op: Operator) -> Result<()> {
-    let path = format!("{} !@#$%^&*()_+-=;'><,?.txt", uuid::Uuid::new_v4());
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-
-    op.object(&path).blocking_write(content)?;
-
-    let meta = op
-        .object(&path)
-        .blocking_metadata()
-        .expect("stat must succeed");
-    assert_eq!(meta.content_length(), size as u64);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Stat existing file should return metadata
-fn test_blocking_stat(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-
-    op.object(&path)
-        .blocking_write(content)
-        .expect("write must succeed");
-
-    let meta = op.object(&path).blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::FILE);
-    assert_eq!(meta.content_length(), size as u64);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Stat existing file should return metadata
-fn test_blocking_stat_dir(op: Operator) -> Result<()> {
-    let path = format!("{}/", uuid::Uuid::new_v4());
-
-    op.object(&path)
-        .blocking_create()
-        .expect("write must succeed");
-
-    let meta = op.object(&path).blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::DIR);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Stat existing file with special chars should return metadata
-fn test_blocking_stat_with_special_chars(op: Operator) -> Result<()> {
-    let path = format!("{} !@#$%^&*()_+-=;'><,?.txt", uuid::Uuid::new_v4());
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-
-    op.object(&path)
-        .blocking_write(content)
-        .expect("write must succeed");
-
-    let meta = op.object(&path).blocking_metadata()?;
-    assert_eq!(meta.mode(), ObjectMode::FILE);
-    assert_eq!(meta.content_length(), size as u64);
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Stat not exist file should return NotFound
-fn test_blocking_stat_not_exist(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-
-    let meta = op.object(&path).blocking_metadata();
-    assert!(meta.is_err());
-    assert_eq!(meta.unwrap_err().kind(), io::ErrorKind::NotFound);
-
-    Ok(())
-}
-
-/// Read full content should match.
-fn test_blocking_read_full(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-
-    op.object(&path)
-        .blocking_write(content.clone())
-        .expect("write must succeed");
-
-    let bs = op.object(&path).blocking_read()?;
-    assert_eq!(size, bs.len(), "read size");
-    assert_eq!(
-        format!("{:x}", Sha256::digest(&bs)),
-        format!("{:x}", Sha256::digest(&content)),
-        "read content"
-    );
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Read range content should match.
-fn test_blocking_read_range(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-    debug!("Generate a random file: {}", &path);
-    let (content, size) = gen_bytes();
-    let (offset, length) = gen_offset_length(size as usize);
-
-    op.object(&path)
-        .blocking_write(content.clone())
-        .expect("write must succeed");
-
-    let bs = op
-        .object(&path)
-        .blocking_range_read(offset..offset + length)?;
-    assert_eq!(bs.len() as u64, length, "read size");
-    assert_eq!(
-        format!("{:x}", Sha256::digest(&bs)),
-        format!(
-            "{:x}",
-            Sha256::digest(&content[offset as usize..(offset + length) as usize])
-        ),
-        "read content"
-    );
-
-    op.object(&path)
-        .blocking_delete()
-        .expect("delete must succeed");
-    Ok(())
-}
-
-/// Read not exist file should return NotFound
-fn test_blocking_read_not_exist(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-
-    let bs = op.object(&path).blocking_read();
-    assert!(bs.is_err());
-    assert_eq!(bs.unwrap_err().kind(), io::ErrorKind::NotFound);
 
     Ok(())
 }
@@ -715,24 +421,6 @@ async fn test_list_dir_with_file_path(op: Operator) -> Result<()> {
     let obs = op.object(&parent).list().await.map(|_| ());
     assert!(obs.is_err());
     assert!(obs.unwrap_err().to_string().contains("Not a directory"));
-
-    Ok(())
-}
-
-// Delete existing file should succeed.
-fn test_blocking_delete(op: Operator) -> Result<()> {
-    let path = uuid::Uuid::new_v4().to_string();
-    debug!("Generate a random file: {}", &path);
-    let (content, _) = gen_bytes();
-
-    op.object(&path)
-        .blocking_write(content)
-        .expect("write must succeed");
-
-    op.object(&path).blocking_delete()?;
-
-    // Stat it again to check.
-    assert!(!op.object(&path).blocking_is_exist()?);
 
     Ok(())
 }
