+++ tests/behavior/main.rs
@@ -21,6 +21,8 @@ mod blocking_read;
 mod utils;
 #[macro_use]
 mod write;
+#[macro_use]
+mod blocking_write;
 
 pub fn init_logger() {
     let _ = env_logger::builder().is_test(true).try_init();
@@ -37,6 +39,8 @@ macro_rules! behavior_tests {
             behavior_blocking_read_tests!($service);
             // can_read && can_write
             behavior_write_tests!($service);
+            // can_read && can_write && can_blocking
+            behavior_blocking_write_tests!($service);
         )*
     };
 }
